LPS-202879 portal-cluster-multiple - moving ClusterChannel and ClusterReceiver to portal-kernel module to allow ClusterExecutorImpl methods to be also moved, prepare for next
# breaking
## What modules/apps/portal/portal-cluster-multiple/src/main/java/com/liferay/portal/cluster/multiple/internal/ClusterChannel.java
modules/apps/portal/portal-cluster-multiple/src/main/java/com/liferay/portal/cluster/multiple/internal/ClusterChannel.java was moved to portal-kernel/src/com/liferay/portal/kernel/cluster/ClusterChannel.java

## Why
This allow ClusterExecutor interface to have all ClusterExecutorImorImpl methods, avoiding the nowne class registered

# breaking
## What modules/apps/portal/portal-cluster-multiple/src/main/java/com/liferay/portal/cluster/multiple/internal/ClusterReceiver.java
modules/apps/portal/portal-cluster-multiple/src/main/java/com/liferay/portal/cluster/multiple/intenral/ClusterReceiver.java was moved to portal-kernel/src/com/liferay/portal/kernel/cluster/ClusterReiver.java

## Why
This allow ClusterExecutor interface to have all ClusterExecutorImorImpl methods, avoiding the nowne class registered

----
