+++ modules/apps/portal/portal-cluster-multiple/src/main/java/com/liferay/portal/cluster/multiple/internal/BaseClusterReceiver.java
@@ -7,6 +7,7 @@ package com.liferay.portal.cluster.multiple.internal;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.cluster.Address;
+import com.liferay.portal.kernel.cluster.ClusterReceiver;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
